+++ lib/doggo/components/icon.ex
@@ -8,16 +8,25 @@ defmodule Doggo.Components.Icon do
   @impl true
   def doc do
     """
-    Renders a customizable icon using a slot for SVG content.
+    Renders an icon with optional text.
 
-    This component does not bind you to a specific set of icons. Instead, it
-    provides a slot for inserting SVG content from any icon library you choose.
+    The component does not make assumptions about the icon library. Instead, it
+    allows you to reference functions from libraries or custom functions that
+    render SVG icons.
     """
   end
 
   @impl true
   def builder_doc do
     """
+    - `:icon_module` (required) - The module that defines the function
+      component(s) for rendering the icon SVG elements.
+    - `:icon_fun` - The name of a function component defined in the icon module
+      that has a `name` attribute (string) renders the corresponding icon
+      SVG element. If not set, the component will use the function component
+      with the same name as the icon name and not set any attributes.
+    - `:names` - Either a list of available icon names or a 0-arity function
+      that returns the list. This is only used in the generated storybook.
     - `:text_position_after_class` - This class is added to the root element if
       `:text_position` is set to `"after"`.
     - `:text_position_before_class` - This class is added to the root element if
@@ -32,19 +41,79 @@ defmodule Doggo.Components.Icon do
   @impl true
   def usage do
     """
-    Render an icon with visually hidden text using the `heroicons` library:
+    ## Configuration
+
+    For icon libraries that define a separate function component for each
+    individual icon such as `heroicons`, you need to set the `icon_module`
+    option.
+
+    ```elixir
+    defmodule MyAppWeb.CoreComponents do
+      use Doggo.Components
+      use Phoenix.Component
+
+      build_icon(icon_module: Heroicons)
+    end
+    ```
+
+    The `name` attribute passed to the generated icon component needs to
+    reference a function component in the configured module.
 
     ```heex
-    <.icon text="report bug"><Heroicons.bug_ant /></.icon>
+    <.icon name="bug_ant" />
+    ```
+
+    In this example, the icon component will use `Heroicons.bug_ant/1` to render
+    the SVG icon in its inner markup.
+
+    Names are internally normalized by replacing dashes with underscores.
+    Therefore, both `name="bug_ant"` and `name="bug-ant"` will work.
+
+    For icon libraries that expose a single function component, you can
+    additionally set the `icon_fun` option.
+
+    ```elixir
+    build_icon(icon_module: MyIcons, icon_fun: :render)
+    ```
+
+    In that case, the generated icon component will pass the `name` attribute
+    on to the referenced function component.
+
+    ```heex
+    <.icon name="circle-question" />
+    ```
+
+    In this example, the generated markup will be similar to:
+
+    ```heex
+    <span class="icon">
+      <MyIcons.render name="circle-question" />
+    </span>
+    ```
+
+    ## Text display
+
+    Render an icon with visually hidden text:
+
+    ```heex
+    <.icon name="bug_ant" text="report bug" />
     ```
 
     To display the text visibly:
 
     ```heex
-    <.icon text="report bug" text_position="after">
-      <Heroicons.bug_ant />
-    </.icon>
+    <.icon name="bug_ant" text="report bug" text_position="after" />
     ```
+
+    Or:
+
+    ```heex
+    <.icon name="bug_ant" text="report bug" text_position="before" />
+    ```
+
+    The `text_position` attribute values are chosen to work with both
+    left-to-right and right-to-left languages. Refer to the CSS example for
+    applying the position correctly.
 
     > #### aria-hidden {: .info}
     >
@@ -66,6 +135,9 @@ defmodule Doggo.Components.Icon do
       maturity: :refining,
       modifiers: [],
       extra: [
+        icon_module: nil,
+        icon_fun: nil,
+        names: [],
         text_position_after_class: "has-text-after",
         text_position_before_class: "has-text-before",
         text_position_hidden_class: nil,
@@ -82,7 +154,7 @@ defmodule Doggo.Components.Icon do
   @impl true
   def attrs_and_slots do
     quote do
-      slot :inner_block, doc: "Slot for the SVG element.", required: true
+      attr :name, :string, required: true, doc: "The name of the icon."
 
       attr :text, :string,
         default: nil,
@@ -104,6 +176,25 @@ defmodule Doggo.Components.Icon do
 
   @impl true
   def init_block(_opts, extra) do
+    icon_module = Keyword.fetch!(extra, :icon_module)
+    icon_fun = Keyword.fetch!(extra, :icon_fun)
+
+    if is_nil(icon_module) do
+      raise """
+      missing icon_module option
+
+      The icon component requires the `icon_module` option to be set.
+
+      If the `icon_module` option is set without the `icon_fun` option, the
+      module is expected to define function components that render the icons
+      with the same name as the `name` attribute passed to the icon component.
+
+      If the `icon_fun` option is set, the function referenced with
+      `icon_module` and `icon_fun` must be a function component that accepts
+      a `name` string attribute and renders the corresponding icon.
+      """
+    end
+
     text_position_after_class =
       Keyword.fetch!(extra, :text_position_after_class)
 
@@ -133,6 +224,8 @@ defmodule Doggo.Components.Icon do
         |> var!()
         |> Map.update!(:class, &(&1 ++ [text_position_class]))
         |> assign(:text_class, text_class)
+        |> assign(:icon_module, unquote(icon_module))
+        |> assign(:icon_fun, unquote(icon_fun))
     end
   end
 
@@ -140,11 +233,36 @@ defmodule Doggo.Components.Icon do
   def render(assigns) do
     ~H"""
     <span class={@class} {@rest}>
-      {render_slot(@inner_block)}
+      <Doggo.Components.Icon.dynamic_icon
+        name={@name}
+        module={@icon_module}
+        fun={@icon_fun}
+      />
       <span :if={@text} class={@text_class}>
         {@text}
       </span>
     </span>
     """
+  end
+
+  @doc false
+
+  attr :name, :string, required: true
+  attr :module, :atom
+  attr :fun, :atom
+
+  def dynamic_icon(%{module: module, fun: nil} = assigns)
+      when is_atom(module) do
+    {module, assigns} = Map.pop!(assigns, :module)
+    {name, assigns} = Map.pop!(assigns, :name)
+    name = String.replace(name, "-", "_")
+    apply(module, String.to_existing_atom(name), [assigns])
+  end
+
+  def dynamic_icon(%{module: module, fun: fun} = assigns)
+      when is_atom(module) and is_atom(fun) do
+    {module, assigns} = Map.pop!(assigns, :module)
+    {fun, assigns} = Map.pop!(assigns, :fun)
+    apply(module, fun, [assigns])
   end
 end